move publishing workspace actions into publishing module
