+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/publish-with-descendants/workspace-action/publish-with-descendants.action.ts
@@ -1,4 +1,4 @@
-import { UMB_DOCUMENT_WORKSPACE_CONTEXT } from '../document-workspace.context-token.js';
+import { UMB_DOCUMENT_WORKSPACE_CONTEXT } from '../../../workspace/document-workspace.context-token.js';
 import { UmbWorkspaceActionBase } from '@umbraco-cms/backoffice/workspace';
 
 export class UmbDocumentPublishWithDescendantsWorkspaceAction extends UmbWorkspaceActionBase {
