+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/schedule-publish/manifests.ts
@@ -1,3 +1,4 @@
 import { manifests as modalManifests } from './modal/manifests.js';
+import { manifests as workspaceActionManifests } from './workspace-action/manifests.js';
 
-export const manifests: Array<UmbExtensionManifest> = [...modalManifests];
+export const manifests: Array<UmbExtensionManifest> = [...modalManifests, ...workspaceActionManifests];
