+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/unpublish/manifests.ts
@@ -1,9 +1,11 @@
 import { manifests as entityActionManifests } from './entity-action/manifests.js';
 import { manifests as entityBulkActionManifests } from './entity-bulk-action/manifests.js';
 import { manifests as modalManifests } from './modal/manifests.js';
+import { manifests as workspaceActionManifests } from './workspace-action/manifests.js';
 
 export const manifests: Array<UmbExtensionManifest> = [
 	...entityActionManifests,
 	...entityBulkActionManifests,
 	...modalManifests,
+	...workspaceActionManifests,
 ];
