+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/unpublish/workspace-action/manifests.ts
@@ -1,0 +1,27 @@
+import { UMB_USER_PERMISSION_DOCUMENT_UNPUBLISH } from '../../../user-permissions/constants.js';
+import { UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS } from '@umbraco-cms/backoffice/recycle-bin';
+
+export const manifests: Array<UmbExtensionManifest> = [
+	{
+		type: 'workspaceActionMenuItem',
+		kind: 'default',
+		alias: 'Umb.Document.WorkspaceActionMenuItem.Unpublish',
+		name: 'Unpublish',
+		weight: 0,
+		api: () => import('./unpublish.action.js'),
+		forWorkspaceActions: 'Umb.WorkspaceAction.Document.SaveAndPublish',
+		meta: {
+			label: '#actions_unpublish',
+			icon: 'icon-globe',
+		},
+		conditions: [
+			{
+				alias: 'Umb.Condition.UserPermission.Document',
+				allOf: [UMB_USER_PERMISSION_DOCUMENT_UNPUBLISH],
+			},
+			{
+				alias: UMB_ENTITY_IS_NOT_TRASHED_CONDITION_ALIAS,
+			},
+		],
+	},
+];
