+++ src/Umbraco.Web.UI.Client/src/packages/documents/documents/publishing/unpublish/workspace-action/unpublish.action.ts
@@ -1,4 +1,4 @@
-import { UMB_DOCUMENT_WORKSPACE_CONTEXT } from '../document-workspace.context-token.js';
+import { UMB_DOCUMENT_WORKSPACE_CONTEXT } from '../../../workspace/constants.js';
 import { UmbWorkspaceActionBase } from '@umbraco-cms/backoffice/workspace';
 
 export class UmbDocumentUnpublishWorkspaceAction extends UmbWorkspaceActionBase {
